Replaced repeated paths in .emacs with constants.
When changing the name of my elisp directory, I got annoyed with all
the places I had to replace the paths to files via query-replace
whatever.  I replaced the long paths with "constants".  Strangely,
this worked out to be actually more code in this circumstance, but if
I ever mess around with my directory structure again I'll be glad I
did this.