Add hook to kill emacsclient buffers with C-x k.
I've been improperly closing emacsclient opened buffers with C-x k.
When I do this, I get an annoying "Buffer still has clients" prompt.
What I'm supposed to do is close an emacsclient buffer with C-x #, the
server-edit command, but that is annoying to remember. This sets C-x k
to server-edit when I'm editing a buffer opened with emacsclient.