Enable auto-revert-mode.
This is especially useful when I'm editing a file with multiple tools.
For example, I may be writing java in Eclipse, but I need to perform
some complex text manipulation so I open up an emacs window to do it.
This mode will automatically keep my emacs buffer up to date with the
changes.