Update flymake python support for my work machine.
The previous flymake-pylint-init function that I copy pasted from the
internet didn't work correctly when I have symlinks in the path. Also,
the name of the command I'm supposed to run on this machine is epylint,
not epylint.py.

This change includes a redefined flymake-get-file-name-mode-and-masks
function that I copied from the internet so I can set up flymake to
start when I enter a mode instead of only when given an appropriate file
extension. I need this for flymake to work on many of the tools comms
scripts that lack the .py extension.