Scroll using space/backspace in evil's normal state.
I have space and backspace set to scroll forward or backward a page in
my .vimrc. The default vim binding for space is the same as l and
backspace is the same as h, so neither space nor backspace are essential
key bindings. I prefer to give scrolling a more ergonomic binding than
C-f/C-b and space/backspace matches emacs' view-mode and info-mode.