Make evil's * and # behave like vim.
Vim treats a word separated by underscores/hyphens as a single word when
searching with * or #. So hitting * on the first part of
THIS_GLOBAL_VARIABLE will search for THIS_GLOBAL_VARIABLE instead of
THIS. By default, evil works more like emacs, which treats each part as
an individual word. While I want this behavior for most commands, like
forward-word, 95% of the time when I use * or # I'm looking for a full
symbol elsewhere in the file.