Add my-speedbar-get-focus to stop unnecessary refreshing.
The standard speedbar-get-focus function refreshes the speedbar, which
will unexpand all the buffers shown in buffer mode. This makes buffer
mode much less useful, so I've created my own version of the function
that doesn't refresh the buffer. I also removed the "g" binding in the
speedbar so I can use gg to go the top of the buffer with evil.