Fix evil syntax table for / and ? commands.
I've made fixes for * and # in the past to allow underscores and hyphens
to be treated as part of a word. This extends that behavior to the
search commands. I should probably figure out a better way to do this
because this is creating a lot of duplicate advice for functions.