Remove advice for evil search commands regarding -/_
This doesn't work in the current version of emacs and I can't figure
out why. I get errors about not finding either the word in question or
some other random character. I'll have to learn to live without it
because I don't want to completely change the syntax table for
everything.